Add Anemometer messages to MAVLink via ardupilotmega.xml, message 188. Tweaks to AP_Anemometer - add sample code for calculating angle. Issue: Used int16 which can't hold >32768 for angle up to 36500; unsigned not possible. Should switch to tenths of a degree instead of hundredths.

+++ libraries/AP_Anemometer/AP_Anemometer_Custom.h
@@ -22,7 +22,7 @@ public:
     /* AP_Anemometer public interface: */
     bool            init();
     uint8_t         read();
-    float           get_anglecd();
+    int16_t         get_anglecd();
     float           get_speed();
 
 private:
