+++ libraries/AP_Anemometer/AP_Anemometer_HIL.h
@@ -15,7 +15,7 @@ private:
 public:
     bool init();
     uint8_t read();
-    float get_anglecd();
+    int16_t get_anglecd();
     float get_speed();
     void setHIL(float angle, float speed);
 };
